Rename mockup tests and remove duplicate
